# perf: reduce allocations in NewV3 and NewV5