```
goos: linux
goarch: amd64
pkg: github.com/gofrs/uuid/v5
cpu: 12th Gen Intel(R) Core(TM) i5-12400F
                   │     old     │                 new                 │
                   │   sec/op    │   sec/op     vs base                │
Generator/NewV3-12   169.5n ± 1%   102.5n ± 1%  -39.55% (p=0.000 n=10)
Generator/NewV5-12   183.2n ± 2%   113.6n ± 1%  -37.96% (p=0.000 n=10)

                   │    old     │                  new                  │
                   │    B/op    │   B/op    vs base                     │
Generator/NewV3-12   144.0 ± 0%   0.0 ± 0%  -100.00% (p=0.000 n=10)
Generator/NewV5-12   168.0 ± 0%   0.0 ± 0%  -100.00% (p=0.000 n=10)

                   │    old     │                   new                   │
                   │ allocs/op  │ allocs/op   vs base                     │
Generator/NewV3-12   4.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
Generator/NewV5-12   4.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
```