@cameracker 2025-09-25
@anatoly-kussul  thank you very much for this contribution, and apologies it took me a while to get to it. Nice change! Simple but impactful perf improvements are appreciated.